Addding order notifications

+++ app/src/main/java/com/bijoyskochar/smartcart/items/OrderInformation.java
@@ -20,7 +20,7 @@ public class OrderInformation extends JsonModel implements Serializable {
     public ChipItem chip;
 
     @JsonField
-    public Integer cart_weight;
+    public Double cart_weight;
 
     @JsonField
     public String created;
